requirements_omicsdmv1.txt
aiofiles==24.1.0
aniso8601==9.0.1
anyio==4.9.0
async-property==0.2.2
attrs==22.1.0
boto3==1.17.88
botocore==1.20.112
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
click==8.1.3
coverage==6.5.0
cryptography==3.4.7
deprecation==2.1.0
flake8==4.0.1
flake8-html==0.4.1
Flask==2.0.3
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==2.8
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.17.0
jwcrypto==1.5.6
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.13.0
PyJWT==2.3.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-keycloak==5.5.1
pytz==2022.6
requests==2.25.1
requests-toolbelt==1.0.0
s3transfer==0.4.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.44
toml==0.10.2
tomli==2.0.1
typing_extensions==4.13.2
urllib3==1.26.5
Werkzeug==2.0.3
zipp==3.10.0
